Improved clarity for instructions and added pretrained weights.

+++ utils/create_datalist.py
@@ -25,7 +25,8 @@ def create_dataset_json(
     labels: str, 
     modality: str,
     image_file_extension: str,
-    label_file_extension: str
+    label_file_extension: str,
+    inference: bool = False
     ) -> list:
     data_json_new = dict()
     
@@ -44,7 +45,8 @@ def create_dataset_json(
     train_list = np.array(list(train_list))
 
     np.random.shuffle(train_list)
-    train_list, test_list = train_list[:int(len(train_list)*0.8)], train_list[int(len(train_list)*0.8):]
+    # train_list, test_list = train_list[:int(len(train_list)*0.8)], train_list[int(len(train_list)*0.8):]
+    test_list = train_list.copy()
 
     data_json_new["training"] = [
         {
@@ -55,31 +57,35 @@ def create_dataset_json(
     ]
     data_json_new["numTraining"] = len(data_json_new["training"])
 
-    # Remove existing files in imagesTs and labelsTs folders
-    shutil.rmtree(dataset_input_dir + "/imagesTs")
-    shutil.rmtree(dataset_input_dir + "/labelsTs")
-    
-    # Recreate empty imagesTs and labelsTs folders
-    os.makedirs(dataset_input_dir + "/imagesTs")
-    os.makedirs(dataset_input_dir + "/labelsTs")
+    if inference:
+        data_json_new["test"] = data_json_new["training"]
+        data_json_new["numTest"] = len(data_json_new["test"])
 
-    # Copy image and label files from imagesTr and labelsTr to imagesTs and labelsTs folders
-    for case in test_list:
-        image_src = os.path.join(dataset_input_dir, "imagesTr", f"{case}{image_file_extension}")
-        image_dst = os.path.join(dataset_input_dir, "imagesTs", f"{case}{image_file_extension}")
-        label_src = os.path.join(dataset_input_dir, "labelsTr", f"{case}{label_file_extension}")
-        label_dst = os.path.join(dataset_input_dir, "labelsTs", f"{case}{label_file_extension}")
-        shutil.copyfile(image_src, image_dst)
-        shutil.copyfile(label_src, label_dst)
+    else:
+        if os.path.exists(dataset_input_dir + "/imagesTs"):
+            shutil.rmtree(dataset_input_dir + "/imagesTs")
+            shutil.rmtree(dataset_input_dir + "/labelsTs")
 
-    data_json_new["test"] = [
-        {
-            "image": f"./imagesTs/{case}{image_file_extension}",
-            "label": f"./labelsTs/{case}{label_file_extension}"
-        }
-        for case in test_list
-    ]
-    data_json_new["numTest"] = len(data_json_new["test"])
+        os.makedirs(dataset_input_dir + "/imagesTs")
+        os.makedirs(dataset_input_dir + "/labelsTs")
+
+        # Copy image and label files from imagesTr and labelsTr to imagesTs and labelsTs folders
+        for case in test_list:
+            image_src = os.path.join(dataset_input_dir, "imagesTr", f"{case}{image_file_extension}")
+            image_dst = os.path.join(dataset_input_dir, "imagesTs", f"{case}{image_file_extension}")
+            label_src = os.path.join(dataset_input_dir, "labelsTr", f"{case}{label_file_extension}")
+            label_dst = os.path.join(dataset_input_dir, "labelsTs", f"{case}{label_file_extension}")
+            shutil.copyfile(image_src, image_dst)
+            shutil.copyfile(label_src, label_dst)
+
+        data_json_new["test"] = [
+            {
+                "image": f"./imagesTs/{case}{image_file_extension}",
+                "label": f"./labelsTs/{case}{label_file_extension}"
+            }
+            for case in test_list
+        ]
+        data_json_new["numTest"] = len(data_json_new["test"])
     
     return data_json_new
 
@@ -96,7 +102,8 @@ def create_datalist(args):
     dataset_new = create_dataset_json(
         args.input_dir + args.task_name, 
         args.task_name, args.description, args.labels,
-        args.modality, image_file_extension, label_file_extension
+        args.modality, image_file_extension, label_file_extension,
+        args.inference
         )
 
     dataset_with_folds = dataset_new.copy()
@@ -121,7 +128,7 @@ def create_datalist(args):
         os.makedirs(args.output_dir)
 
     with open(os.path.join(args.output_dir, f"dataset_task{args.task_name.split('_')[0][-2:]}_f0.json"), "w") as f:
-        json.dump(dataset_with_folds, f)
+        json.dump(dataset_with_folds, f, indent=2)
         print(f"data list fold_0 for {args.task_name} has been created!")
         f.close()
 
@@ -131,13 +138,15 @@ if __name__ == "__main__":
     
     # change these with info for the task you want to update
     parser.add_argument("-input_dir", "--input_dir", type=str, 
-                        default="/mnt/data/Experiment/Data/MorphiNet-MR_CT/")
+                        # default="/mnt/data/Experiment/Data/MorphiNet-MR_CT/")
+                        default="/mnt/data/Experiment/nnUNet/nnUNet_raw/")
     parser.add_argument("-file_extension", "--file_extension", type=str, 
                         default=".nii.gz", help="the file extension of the data (.nii.gz / .nrrd)")
     parser.add_argument("-task_name", "--task_name", type=str, 
-                        default="Dataset020_SCOTHEART", help="the task name")
+                        # default="Dataset010_CAP_SAX_NRRD", help="the task name")
+                        default="Dataset022_MMWHS_CT", help="the task name")
     parser.add_argument("-d", "--description", help="the task description",
-                        default="CT image data w/ cross-validation")
+                        default="MMWHS CT image data w/ cross-validation")
     parser.add_argument("-l", "--labels", type=json.loads, help="the label name",
                         default='{"0": "background", "1": "lv", "2": "lv-myo", "3": "rv", "4": "rv-myo"}')
     parser.add_argument("-m", "--modality", type=str, 
@@ -149,6 +158,7 @@ if __name__ == "__main__":
                         default=5, help="number of folds")
     parser.add_argument("-seed", "--seed", type=int, 
                         default=42, help="seed number")
+    parser.add_argument("-infer", "--inference", action="store_true", help="whether to create the json list solely for inference.")
 
     args = parser.parse_args()
